requirements.txt
crewai>=0.40.0
crewai-tools>=0.4.0
langchain-openai>=0.1.0
tavily-python>=0.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6